+++ packages/backend/src/controllers/authentication/email/login.ts
@@ -1,4 +1,4 @@
-import { ERRORS_LOGIN, FeatureFlagName } from "@vao/shared-bridge";
+import { ERRORS_LOGIN } from "@vao/shared-bridge";
 import type { NextFunction, Response } from "express";
 
 import { config } from "../../../config";
@@ -6,14 +6,13 @@ import { schema } from "../../../helpers/schema";
 import { status } from "../../../helpers/users";
 import Session from "../../../services/common/Session";
 import CommonUser from "../../../services/common/Users";
-import { FeatureFlagService } from "../../../services/featureFlagService";
 import User from "../../../services/User";
 import { UserRequest } from "../../../types/request";
 import { UsersService } from "../../../usagers/users/users.service";
 import AppError from "../../../utils/error";
 import { logger } from "../../../utils/logger";
-import { OtpService } from "../../../utils/otp";
 import { signAccessToken, signRefreshToken } from "../../../utils/token";
+import { checkActionsOtp } from "../../common/authentication/email/login";
 
 const log = logger(module.filename);
 
@@ -106,26 +105,12 @@ export default async function login(
         ),
       );
     }
-
-    const featureFlags = await FeatureFlagService.getFeatureFlagsAvailable();
-    const requires2FA = await FeatureFlagService.isFeatureAvailable(
-      FeatureFlagName.AUTH_2FA,
-    );
-    if (requires2FA) {
-      const { isLocked } = OtpService.isLocked({
-        otpAttempts: user.otpAttempts ?? 0,
-        otpAttemptsAt: user.otpAttemptsAt ?? null,
+    const { featureFlags, isUpdateOtpNecessary, requires2FA } =
+      await checkActionsOtp({ user });
+    if (isUpdateOtpNecessary) {
+      await UsersService.updateOtpCode({
+        userId: Number(user.id),
       });
-      const isExpired = OtpService.isExpired({
-        otpCodeExpiresAt: user.otpCodeExpiresAt ?? null,
-      });
-      if (!isLocked) {
-        if (isExpired) {
-          await UsersService.updateOtpCode({
-            userId: Number(user.id),
-          });
-        }
-      }
     } else {
       const accessToken = signAccessToken(user);
       const refreshToken = signRefreshToken(user);
