+++ packages/backend/src/controllers/bo-authentication/email/login.ts
@@ -1,4 +1,4 @@
-import { ERRORS_LOGIN, FeatureFlagName, UserDto } from "@vao/shared-bridge";
+import { ERRORS_LOGIN, UserDto } from "@vao/shared-bridge";
 import type { NextFunction, Response } from "express";
 
 import { UsersService } from "../../../admin/users/users.service";
@@ -8,12 +8,11 @@ import { status } from "../../../helpers/users";
 import User from "../../../services/BoUser";
 import Session from "../../../services/common/Session";
 import CommonUser from "../../../services/common/Users";
-import { FeatureFlagService } from "../../../services/featureFlagService";
 import { UserRequest } from "../../../types/request";
 import { signAccessToken, signRefreshToken } from "../../../utils/bo-token";
 import AppError from "../../../utils/error";
 import { logger } from "../../../utils/logger";
-import { OtpService } from "../../../utils/otp";
+import { checkActionsOtp } from "../../common/authentication/email/login";
 
 const log = logger(module.filename);
 
@@ -92,25 +91,12 @@ export default async function login(
       roles: user.roles ?? [],
     };
 
-    const featureFlags = await FeatureFlagService.getFeatureFlagsAvailable();
-    const requires2FA = await FeatureFlagService.isFeatureAvailable(
-      FeatureFlagName.AUTH_2FA,
-    );
-    if (requires2FA) {
-      const { isLocked } = OtpService.isLocked({
-        otpAttempts: user.otpAttempts ?? 0,
-        otpAttemptsAt: user.otpAttemptsAt ?? null,
+    const { featureFlags, isUpdateOtpNecessary, requires2FA } =
+      await checkActionsOtp({ user });
+    if (isUpdateOtpNecessary) {
+      await UsersService.updateOtpCode({
+        userId: Number(user.id),
       });
-      const isExpired = OtpService.isExpired({
-        otpCodeExpiresAt: user.otpCodeExpiresAt ?? null,
-      });
-      if (!isLocked) {
-        if (isExpired) {
-          await UsersService.updateOtpCode({
-            userId: Number(user.id),
-          });
-        }
-      }
     } else {
       const accessToken = signAccessToken(payload);
       const refreshToken = signRefreshToken(payload);
